feat(nui): convert the loading screen to NUI
