+++ engine/src/main/java/org/destinationsol/SolApplication.java
@@ -360,6 +360,7 @@ public class SolApplication implements ApplicationListener {
         }
 
         factionDisplay = new FactionDisplay(context.get(SolCam.class));
+        nuiManager.removeScreen(menuScreens.loading);
         inputManager.setScreen(this, solGame.getScreens().mainGameScreen);
     }
