--- engine/src/main/java/org/destinationsol/menu/LoadingScreen.java
@@ -1,68 +1,0 @@
-/*
- * Copyright 2018 MovingBlocks
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.destinationsol.menu;
-
-import com.badlogic.gdx.graphics.Texture;
-import com.badlogic.gdx.graphics.g2d.TextureAtlas;
-import org.destinationsol.SolApplication;
-import org.destinationsol.assets.Assets;
-import org.destinationsol.common.SolColor;
-import org.destinationsol.game.WorldConfig;
-import org.destinationsol.ui.DisplayDimensions;
-import org.destinationsol.ui.FontSize;
-import org.destinationsol.ui.SolInputManager;
-import org.destinationsol.ui.SolUiBaseScreen;
-import org.destinationsol.ui.UiDrawer;
-
-public class LoadingScreen extends SolUiBaseScreen {
-    private DisplayDimensions displayDimensions;
-
-    private final TextureAtlas.AtlasRegion backgroundTexture;
-
-    private boolean loadTutorial;
-    private boolean isNewGame;
-    private String shipName;
-    private WorldConfig worldConfig;
-
-    LoadingScreen() {
-        displayDimensions = SolApplication.displayDimensions;
-
-        backgroundTexture = Assets.getAtlasRegion("engine:mainMenuBg", Texture.TextureFilter.Linear);
-    }
-
-    @Override
-    public void updateCustom(SolApplication solApplication, SolInputManager.InputPointer[] inputPointers, boolean clickedOutside) {
-        solApplication.play(loadTutorial, shipName, isNewGame, worldConfig);
-    }
-
-    @Override
-    public void drawText(UiDrawer uiDrawer, SolApplication solApplication) {
-        uiDrawer.drawString("Loading...", displayDimensions.getRatio() / 2, .5f, FontSize.MENU, true, SolColor.WHITE);
-    }
-
-    public void setMode(boolean loadTutorial, String shipName, boolean isNewGame, WorldConfig worldConfig) {
-        this.loadTutorial = loadTutorial;
-        this.shipName = shipName;
-        this.isNewGame = isNewGame;
-        this.worldConfig = worldConfig;
-    }
-
-    @Override
-    public void drawBackground(UiDrawer uiDrawer, SolApplication solApplication) {
-        uiDrawer.draw(backgroundTexture, displayDimensions.getRatio(), 1, displayDimensions.getRatio() / 2, 0.5f, displayDimensions.getRatio() / 2, 0.5f, 0, SolColor.WHITE);
-    }
-}
