+++ engine/src/main/resources/org/destinationsol/assets/skins/mainMenu.skin
@@ -30,6 +30,9 @@
         },
         "creditsButton": {
             "font": "engine:main#0.60"
+        },
+        "loadingScreenText": {
+            "font": "engine:main#1.0"
         }
     }
 }
