+++ engine/src/main/resources/org/destinationsol/assets/ui/mainMenu/loadingScreen.ui
@@ -1,0 +1,20 @@
+{
+    "type": "LoadingScreen",
+    "skin": "engine:mainMenu",
+    "contents": {
+        "type": "RelativeLayout",
+        "contents": [
+            {
+                "type": "UILabel",
+                "family": "loadingScreenText",
+                "text": "Loading...",
+                "layoutInfo": {
+                    "position-horizontal-center": {},
+                    "position-vertical-center": {},
+                    "use-content-width": true,
+                    "use-content-height": true
+                }
+            }
+        ]
+    }
+}
